requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backports.zoneinfo==0.2.1
binance-connector-python==0.6.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
dateparser==1.1.8
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.18.0
pyOpenSSL==22.1.0
pyTelegramBotAPI==4.7.1
python-binance==1.0.17
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
regex==2023.5.5
requests==2.28.1
service-identity==21.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
TA-Lib @ file:///home/conda/feedstock_root/build_artifacts/ta-lib_1637266819397/work
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzlocal==5.0.1
ujson==5.7.0
urllib3==1.26.12
websockets==11.0.3
yarl==1.9.2
zope.interface==5.5.0
